irm/api/requirements.txt
#google-cloud-irm==0.1.0
google-api-core[grpc] >= 1.6.0, < 2.0.0dev
google-cloud-monitoring >= 0.31.1, < 0.32dev
